Fix incorrect caching of serialized objects in smart mode
If we smart copy object which has nested indirect object, this nested object will be also serialized and it's content appended to the parent object content. Although when we cached this nested object's content, we cached the hash of the serialized content. This resulted in the situation, that if this nested indirect object was encountered again, object's content hash have been appended to the parent and not the content itself, like it was for the first occurrence.

With this fix some bigger memory overhead is expected for the smart mode. However currently on the real documents that we tested this didn't affect performance, so for now this should suffice. If any performance problems will appear and for more information refer to the DEVSIX-726.

Autoported commit.
Original commit hash: [8bb49d3]