Made sure the null check from a fix does not change execution flow.
The additional null check had been added to the condition of an existing
if clause. This makes the program flow change in case of a null,
suddenly the else if condition is evaluated which is not desired and may
even result in a different exception. Thus, pulled the null check inside
the if block as a separate if clause.

Related to checkin: Fix NullPointerException in drawnig borders logic
for table renderer.

DEVSIX-765
Autoported commit.
Original commit hash: [2b450c7]