Fix for ClassCastException when using PdfA
The test in this commit shows the class cast exception when using background color with PdfA. The change to AbstractRenderer shows a "fix", however I really doubt this is a correct fix for this issue.

Test to show problem with background color in PdfA

Fix to abstract renderer for class cast exception... Note I do not believe this is the correct fix :)

Autoported commit.
Original commit hash: [0a2debb]