Fix processing of newline characters during text rendering
A partial layout status with null overflow renderer used to be returned, which is not correct from the ideological point of view. Now full layout result can be returned in case the text end with a newline and it is the parent's renderer responsibility to handle that renderer correctly

Autoported commit.
Original commit hash: [4c79504]