Process borders correctly while FORCED_PLACEMENT handling
Collapse table borders with cells with big rowspan.
If top border is NO_BORDER do not delete TOP_BORDER property. This can happen
because the cell was splitted (and we should process it) or the cell top border
is NO_BORDER (the behavior doesn't change)

Autoported commit.
Original commit hash: [297b6cc1e9]