Get rid of the default RegexOptions.Compiled option for Regex instance creating
Approximately, it considered that Compiled option reduces regex matching time by ~30%, however it increases the initialization by ~2 times. Therefore, it should only be used in case if this Regex instance is reused heavily, and should not be used in general case. Specifically, this change considerably improves the performance on TypographyIText7Test profiling test.

DEVSIX-1018