Apply tag attributes to L (not List) and TH (not just TD)
BlockRenderer.draw should check for L, not List and for TH as well as
TD when deciding whether to write out attributes. In any case,
attributes are only written out when there are any, so the test now
creates a List and a TH that does write out its attributes (by adding a
ListNumberingType and a colspan).

I think this test also shows a separate (unfixed) bug -- when cells are
added to a table header, they show have the TH, not TD, role.

Autoported commit.
Original commit hash: [bb542f3]