Fix various issues with processing of content streams in case they are not yet read or are just copied
If a stream is just read it's outputStream field won't be initilized which was not handled in PdfCanvas constructor. Also, if the stream is just copied to the document, it will preserve the original encoding (Filter) which also wasn't handled in several places before.

DEVSIX-1082

Autoported commit.
Original commit hash: [737efec280]