Update Put and Get extension methods for the Dictionary class
Actually, because of the way Sharpen is configured, we don't have dictionaries with non-nullable values. Therefore Get and Put methods could be used as they are in java. The only difference is in possiblity in java to put null as keys, while in .NET it is forbidden for the Dictionary class to have null as key for any key types. So I'm not checking for null in put-methods in order to explicitly follow .NET convention and fail early rather than in unpredictable place or way.