Improve leading calculation for consequent lines with different heights
Now layout behaves more like a line-stacking algorithm than a leading-only algorithm. Leading is applied partially as the top spacing before the line and partially as the bottom spacing instead of just being top spacing as before. For consequent lines with different heights this makes a lot of difference in a sense that lines will not overlap

DEVSIX-1053

Autoported commit.
Original commit hash: [484fa88]