Fix text extraction for simple fonts with custom /ToUnicode
In case /Differences have mapping to glyph names that are not present in Adobe Glyph list,
/ToUnicode should be provided to facilitate correct text extraction.
Considering /ToUnicode had a bug in it and now it is fixed.
Also, /ToUnicode mappings to more than one character are now ignored because
codeToUnicode is a simple mapping to one value and in case of ligature mapping
to two characters they used to be glued to one int, resulting in an incorrect
text representation with huge Unicode character ids in it

Autoported commit.
Original commit hash: [560ca87a9]