Add new util to fix inconsistences in exception handling
In C# out of bounds exception for lists and arrays are different
ones, without convenient common parent. So to make tests that rely
on catching then autoported, general exception has to be cached and
checked if it is either of them.

DEVSIX-1557

Autoported commit.
Original commit hash: [61c9f92bd]