Make array return types consistent in PdfShading
The PdfArray now is returned in all cases in PdfShading since it is
the way most other wrappers behaves in similar cases and since it is
easy to get native java array of needed type with PdfArray methods.
The toBooleanArray was added to PdfArray support that usecase.
Also fix some setters not setting Modified flag.

DEVSIX-747

Autoported commit.
Original commit hash: [1249f096a]