Fix exceptions in .NET Standard projects other than .NET Core
The System.Runtime.Loader is supported only for .NET Core despite
having .NET Standard target, and was causing
TypeInitializationExceptions. All usage of this class was moved to
separate util class so this exception can be caught properly.
Also it seems that DependencyContext.Default is returning null for
other than .NET Core projects so null checks were added as well.

DEVSIX-1726