Fix append mode for hybrid-reference documents
iText used to refer to regular xref section by /Prev key in the xref
stream which is "not meaningful in hybrid-reference files"
according to the specification. In full compression mode after the
changes in this commit first an xref stream is written and then a regular
xref section referring to the xref stream. For safety, regular objects,
not the ones from object streams, are also enumerated in the last regular
xref section of the stamped document.

DEVSIX-2080

Autoported commit.
Original commit hash: [68411e67b]