Generalize ResourceResolver and make it easier to subclass
A new method was added for retrieving images that can deal with
FormXObjects, not only with ImageXObjects. Because of breaking
compatibility limitations, this is a new method but in the next
major update of iText it will be removed.
ResourceResolver is now easier to subclass to support other types
of resources that cannot be supported at SXP (Core) level.
Several log message constants have been deprecated because they are
unused

DEVSIX-2096

Autoported commit.
Original commit hash: [3ac6d306a]