Reinitialize form field appearance stream object on every field regeneration
First, appearance stream objects might be shared between different forms fields or other entities, if they are identical, so on regeneration we should not affect those other entities.

Second, the resources of the streams and other entries in stream dictionaries are not shared between regenerations. It seems that this makes sense, because resources are completely rebuilt every time, and all other important entries in the appearance stream are updated on regeneration as well.

DEVSIX-2374

Autoported commit.
Original commit hash: [70def8325]