Fix mark to ligature positioning in GposLookupType5
Previously, when mark was attached to a ligature, the corresponding
glyph in the glyph line was duplicated.
Additionally, we now attach mark to the last applicable component in the
ligature instead of attaching it to the first applicable component. This
is still not the full implementation that keep track of glyph reorderings
and substitutions, but this is a better default behavior

DEVSIX-3726
DEVSIX-2400
DEVSIX-2672
DEVSIX-2713
DEVSIX-3643

Autoported commit.
Original commit hash: [88cda2643]