Disallow removing flushed pages for tagged documents or documents with acroforms
Removing a page triggers a series of accompanying actions, in particular
removing the associated tag structure entries and form field widgets.
When page is flushed its tags are also flushed by default, so this means
that structure subtree(s) corresponding to the flushed pages have already
been flushed when the page is flushed. Removing a page means the subtree
corresponding to the removed page needs to be removed from the tag structure,
and determining the exact subtree is impossible when the tags have been
flushed already. Additionally, for removing the widgets, the correct and
easy way of removing the associated widgets is looking into /Annots array
which is not available when the page is already flushed. In the future
removing pages with associated widgets might become possible if we provide
a connector to forms module which will be respondible for traversing the
form fields and cleaning up necessary widgets, but for now as this works
incorrectly we are disabling this functionality

DEVSIX-4031

Autoported commit.
Original commit hash: [1e6cbc552]