Remove TODOs in Lab and IccBased colors
It seems that these TODOs were added in accordance with the note about the Range entry in Table 65 of the PDF specification 32000-1: 2008.
But it seems that this note is relevant to readers. Thus, we may not take this fact into account when writing to a PDF file.

DEVSIX-4013

Autoported commit.
Original commit hash: [2cf618d25]