Simplify ghostscript util test for -dSAFER command
With the simple test that only checks that malicious code is not executed
we can theoretically have false positive result if command itself
will not work. The idea is that we run more simplistic commands which are
expected to work the same way in more environments.

QA-12194

Autoported commit.
Original commit hash: [dd6071c3a]