Optimize jsoup buffer test
There is no need in checking the issue for strings of length equal to 12, 13, ..., n
We can check it for a string of length m, where m - is the length for which
the issue can be reproduced.

The change has been introduced in a view of a test failing on Android (30+ versions)

DEVSIX-6565

Autoported commit.
Original commit hash: [01507c54e]