Genericise PdfNameTree logic
The new `GenericNameTree` superclass implements most of the
serialisation and deserialisation logic while opening up the possibility
of using name trees that don't live in the document catalog.

The `GenericNameTree` superclass includes a version of buildTree() that
was refactored to use recursion for readability. The benefits of using
iteration over recursion are negligible in this case, since the maximal
depth of the tree is ceil(log(MAX_INT, NODE_SIZE)), which for
NODE_SIZE = 40 is only 6.

iterateItems() was similarly refactored and commented for
maintainability.

getNames() now also returns a PdfString-based map. This is a breaking
change with a ripple effect in various other internal APIs.
To accommodate this change, an accessor interface was added.

RES-629 DEVSIX-2226

Autoported commit.
Original commit hash: [20f53048a]